Refactor OpenAI and OrcaRouter to share OpenAICompatibleClient base
Extract the OpenAI Chat Completions client implementation shared by OpenAI
and OrcaRouter into a new abstract OpenAICompatibleClient base class.
Subclasses now only provide provider-specific defaults (model, endpoint,
API key), which addresses the review question of why OrcaRouter needs a
separate implementation.
